Move motif area related code into v1 directory
REDMINE-19888
