+++ entry_types/scrolled/package/spec/frontend/v1/useMotifAreaState-spec.js
@@ -1,4 +1,4 @@
-import {useMotifAreaState} from 'frontend/useMotifAreaState';
+import {useMotifAreaState} from 'frontend/v1/useMotifAreaState';
 
 import {renderHook, act} from '@testing-library/react-hooks';
 import {fireEvent} from '@testing-library/dom'
