+++ entry_types/scrolled/package/src/frontend/MotifAreaVisibilityProvider.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export const MotifAreaVisibilityContext = React.createContext(false);
+
+export function MotifAreaVisibilityProvider({visible, children}) {
+  return (
+    <MotifAreaVisibilityContext.Provider value={visible}>
+      {children}
+    </MotifAreaVisibilityContext.Provider>
+  );
+}
