+++ entry_types/scrolled/package/src/frontend/v1/Backdrop/BackgroundImage.js
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import {Image} from './../Image';
-import {MotifArea} from './../MotifArea';
-import {useSectionLifecycle} from './../useSectionLifecycle';
-import {Effects} from './Effects';
+import {Image} from '../../Image';
+import {MotifArea} from '../MotifArea';
+import {useSectionLifecycle} from '../../useSectionLifecycle';
+import {Effects} from '../../Backdrop/Effects';
 
 export function BackgroundImage({image, onMotifAreaUpdate, containerDimension}) {
   const {shouldLoad} = useSectionLifecycle();
