+++ entry_types/scrolled/package/src/frontend/v1/index.js
@@ -1,0 +1,2 @@
+export {Backdrop} from './Backdrop';
+export {useMotifAreaState} from './useMotifAreaState';
